Move code to @/lib/invoice
