+++ components/form.js
@@ -18,7 +18,7 @@ import { gql, useLazyQuery } from '@apollo/client'
 import { USER_SUGGESTIONS } from '@/fragments/users'
 import TextareaAutosize from 'react-textarea-autosize'
 import { useToast } from './toast'
-import { useInvoiceable } from './invoice'
+import { useInvoiceable } from '@/lib/invoice'
 import { numWithUnits } from '@/lib/format'
 import textAreaCaret from 'textarea-caret'
 import ReactDatePicker from 'react-datepicker'
