+++ components/item-act.js
@@ -6,7 +6,7 @@ import { useMe } from './me'
 import UpBolt from '@/svgs/bolt.svg'
 import { amountSchema } from '@/lib/validate'
 import { gql, useApolloClient, useMutation } from '@apollo/client'
-import { payOrLoginError, useInvoiceModal } from './invoice'
+import { payOrLoginError, useInvoiceModal } from '@/lib/invoice'
 import { TOAST_DEFAULT_DELAY_MS, useToast, withToastFlow } from './toast'
 import { useLightning } from './lightning'
 import { nextTip } from './upvote'
